Split modules inside lib.rs into their own files
This should make the project easier to manage as it grows.
